+++ src/main.rs
@@ -49,12 +49,32 @@ fn main() -> anyhow::Result<()> {
 
     #[cfg(not(debug_assertions))]
     let (file_log_layer, _f_guard) = {
+        use std::time::{Duration, SystemTime};
+
         use anyhow::Context;
 
         let logs_dir = dirs::data_dir()
             .context("Failed to get $data_dir path")?
             .join("komorebi-switcher")
             .join("logs");
+
+        const MONTH: Duration = Duration::from_secs(60 * 60 * 24 * 30);
+        let now = SystemTime::now();
+
+        // remove logs older than a month
+        for entry in std::fs::read_dir(&logs_dir).ok().into_iter().flatten() {
+            let Ok(entry) = entry else {
+                continue;
+            };
+            let Ok(modified_time) = entry.metadata().and_then(|m| m.modified()) else {
+                continue;
+            };
+
+            let elapsed = now.duration_since(modified_time).unwrap_or_default();
+            if elapsed > MONTH {
+                let _ = std::fs::remove_file(entry.path());
+            }
+        }
 
         let appender = tracing_appender::rolling::daily(&logs_dir, "komorebi-switcher.log");
         let (non_blocking, _guard) = tracing_appender::non_blocking(appender);